[bug] Revert "[bug] Switched to ints for edges"
This reverts commit 0ef5277cd014ff4240fa51c986da0ca6b656c283.

+++ code/mesh/inc/ClosingFacetComputer.hpp
@@ -20,7 +20,7 @@
 class ClosingFacetComputer
 {
     public:
-        typedef std::pair<int,int> Edge;
+        typedef std::pair<size_t,size_t> Edge;
         typedef std::vector<Edge> ListOfEdges;
         ClosingFacetComputer(const Eigen::Matrix3Xd& mesh, const ListOfEdges& edges);
